Bugfix install procedure. Remove bad \! from printf
that shell tries to expand. Also `type' doesn't seem to work inside a
function scope for some reason. So use grep instead, on the tempfile
of aliases that we construct from the existing alias file for flatpak
packages. Plus, check for installed binary name, so we skip flatpak if
the software was installed via package manager.